Change LigneFeuillePrésence frais to double

+++ SystemeSuiviProjets.Core/Entities/LigneFeuillePrésence.cs
@@ -2,11 +2,11 @@
 
 namespace SystemeSuiviProjets.Core
 {
-    public class LigneFeuillePrésence(TypeTarification typeTarification, int nbUnités, decimal fraisDéplacement = 0, decimal fraisSéjour = 0) : BaseEntity
+    public class LigneFeuillePrésence(TypeTarification typeTarification, int nbUnités, double fraisDéplacement = 0, double fraisSéjour = 0) : BaseEntity
     {
         public TypeTarification TypeTarification { get; set; } = typeTarification;
         public int NombreUnités { get; set; } = nbUnités;
-        public decimal FraisDéplacement { get; set; } = fraisDéplacement;
-        public decimal FraisSéjour { get; set; } = fraisSéjour;
+        public double FraisDéplacement { get; set; } = fraisDéplacement;
+        public double FraisSéjour { get; set; } = fraisSéjour;
     }
 }
